合并并精简冗余模块
