+++ CI/unit_tests/providers/test_wjx_reverse_fill.py
@@ -8,7 +8,7 @@ from software.core.questions.schema import QuestionEntry
 from software.core.reverse_fill.schema import REVERSE_FILL_FORMAT_WJX_SCORE, REVERSE_FILL_FORMAT_WJX_SEQUENCE, REVERSE_FILL_FORMAT_WJX_TEXT, REVERSE_FILL_STATUS_BLOCKED, REVERSE_FILL_STATUS_FALLBACK, REVERSE_FILL_STATUS_REVERSE
 from software.core.reverse_fill.validation import build_enabled_reverse_fill_spec, build_reverse_fill_spec
 from software.io.config import RuntimeConfig
-from software.io.spreadsheets import load_wjx_excel_export
+from software.io.spreadsheets.wjx_excel import load_wjx_excel_export
 
 def _write_workbook(rows: list[list[object]]) -> str:
     workbook = Workbook()
