--- software/io/markdown/__init__.py
@@ -1,6 +1,0 @@
-"""Markdown 工具。"""
-
-from software.io.markdown.utils import strip_markdown
-
-__all__ = ["strip_markdown"]
-
