--- software/io/qr/__init__.py
@@ -1,6 +1,0 @@
-"""二维码工具。"""
-
-from software.io.qr.utils import decode_qrcode
-
-__all__ = ["decode_qrcode"]
-
