--- software/io/reports/__init__.py
@@ -1,6 +1,0 @@
-"""报表输出。"""
-
-from software.io.reports.ip_usage_log import get_usage_summary
-
-__all__ = ["get_usage_summary"]
-
