--- software/io/spreadsheets/__init__.py
@@ -1,6 +1,0 @@
-"""Spreadsheet readers."""
-
-from software.io.spreadsheets.wjx_excel import load_wjx_excel_export
-
-__all__ = ["load_wjx_excel_export"]
-
