--- software/system/__init__.py
@@ -1,6 +1,0 @@
-"""系统能力包。"""
-
-from .power_management import SystemSleepBlocker
-
-__all__ = ["SystemSleepBlocker"]
-
