--- software/ui/controller/__init__.py
@@ -1,12 +1,0 @@
-"""UI 控制器包。"""
-
-from software.ui.controller.engine_adapter import BoolVar, EngineGuiAdapter
-from software.ui.controller.run_controller import RunController
-
-__all__ = [
-    "BoolVar",
-    "EngineGuiAdapter",
-    "RunController",
-]
-
-
