--- software/ui/controller/run_controller_parts/__init__.py
@@ -1,12 +1,0 @@
-"""RunController 拆分模块。"""
-
-from .parsing import RunControllerParsingMixin
-from .runtime import RunControllerRuntimeMixin
-from .persistence import RunControllerPersistenceMixin
-
-__all__ = [
-    "RunControllerParsingMixin",
-    "RunControllerRuntimeMixin",
-    "RunControllerPersistenceMixin",
-]
-
