--- software/ui/pages/workbench/dashboard/__init__.py
@@ -1,6 +1,0 @@
-"""工作台概览页。"""
-
-from software.ui.pages.workbench.dashboard.page import DashboardPage
-
-__all__ = ["DashboardPage"]
-
