--- software/ui/pages/workbench/log_panel/__init__.py
@@ -1,6 +1,0 @@
-"""工作台日志页。"""
-
-from software.ui.pages.workbench.log_panel.page import LogPage
-
-__all__ = ["LogPage"]
-
