--- software/ui/pages/workbench/question_editor/__init__.py
@@ -1,1 +1,0 @@
-"""题目配置页面包。"""
