--- software/ui/pages/workbench/reverse_fill/__init__.py
@@ -1,5 +1,0 @@
-"""问卷反填页。"""
-
-from software.ui.pages.workbench.reverse_fill.page import ReverseFillPage
-
-__all__ = ["ReverseFillPage"]
