--- software/ui/pages/workbench/runtime_panel/__init__.py
@@ -1,5 +1,0 @@
-"""运行时界面模块"""
-from software.ui.pages.workbench.runtime_panel.main import RuntimePage
-
-__all__ = ["RuntimePage"]
-
