--- software/ui/pages/workbench/shared/__init__.py
@@ -1,11 +1,0 @@
-"""Workbench shared UI building blocks."""
-
-from .clipboard import SurveyClipboardMixin
-from .random_ip_toggle_row import RandomIpToggleRow
-from .survey_entry_card import SurveyEntryCard
-
-__all__ = [
-    "SurveyClipboardMixin",
-    "RandomIpToggleRow",
-    "SurveyEntryCard",
-]
