--- software/ui/pages/workbench/strategy/__init__.py
@@ -1,5 +1,0 @@
-"""题目策略页导出。"""
-
-from .page import QuestionStrategyPage
-
-__all__ = ["QuestionStrategyPage"]
