--- software/ui/widgets/contact_form/__init__.py
@@ -1,6 +1,0 @@
-"""联系表单组件。"""
-
-from software.ui.widgets.contact_form.widget import ContactForm
-
-__all__ = ["ContactForm"]
-
