+++ wjx/provider/navigation.py
@@ -1,6 +1,5 @@
 """页面导航 - 翻页、跳转与滚动控制"""
 import math
-import random
 import threading
 import time
 from typing import Optional
@@ -9,6 +8,7 @@ from software.core.questions.utils import (
     extract_text_from_element as _extract_text_from_element,
     smooth_scroll_to_element as _smooth_scroll_to_element,
 )
+from software.core.engine.navigation import _human_scroll_after_question
 from software.network.browser import By, BrowserDriver
 from software.logging.log_utils import log_suppressed_exception
 
@@ -385,21 +385,6 @@ def dismiss_resume_dialog_if_present(
             else:
                 time.sleep(poll_interval)
     return False
-
-
-def _human_scroll_after_question(driver: BrowserDriver) -> None:
-    distance = random.uniform(120, 260)
-    page = getattr(driver, "page", None)
-    if page:
-        try:
-            page.mouse.wheel(0, distance)
-            return
-        except Exception as exc:
-            log_suppressed_exception("navigation._human_scroll_after_question mouse wheel", exc)
-    try:
-        driver.execute_script("window.scrollBy(0, arguments[0]);", distance)
-    except Exception as exc:
-        log_suppressed_exception("navigation._human_scroll_after_question script", exc)
 
 
 def _click_next_page_button(driver: BrowserDriver) -> bool:
